move zeek logic to zeek.py
