+++ src/navv/utilities.py
@@ -5,7 +5,6 @@
 import os
 import contextlib
 import json
-from subprocess import Popen, PIPE, STDOUT, check_call
 import time
 
 from netaddr import EUI, core as netaddr_core
@@ -43,27 +42,6 @@ def timeit(method):
         return result
 
     return timed
-
-
-@timeit
-def run_zeek(pcap_path, zeek_logs_path, **kwargs):
-    with pushd(zeek_logs_path):
-        # can we add Site::local_nets to the zeek call here?
-        err = check_call(["zeek", "-C", "-r", pcap_path, "local.zeek"])
-        error_msg(f"Zeek returned with code: {err}")
-
-
-def perform_zeekcut(fields, log_file):
-    """Perform the call to zeek-cut with the identified fields on the specified log file"""
-    try:
-        with open(log_file, "rb") as f:
-            zeekcut = Popen(
-                ["zeek-cut"] + fields, stdout=PIPE, stdin=PIPE, stderr=STDOUT
-            )
-            return zeekcut.communicate(input=f.read())[0]
-    except OSError as e:
-        # probably "file does not exist"
-        return b""
 
 
 def trim_dns_data(data):
